Separate sms notification utils for mocking
