+++ packages/workflow/test/setupJest.ts
@@ -11,7 +11,19 @@
  */
 import { join } from 'path'
 import * as fetch from 'jest-fetch-mock'
+import {
+  informantSMSNotificationMock
+} from '@workflow/test/utils'
 
+jest.mock('@workflow/features/registration/smsNotificationUtils', () => {
+  const actual = jest.requireActual('@workflow/features/registration/smsNotificationUtils')
+  return {
+    ...actual,
+    getInformantSMSNotification: jest
+      .fn()
+      .mockReturnValue(informantSMSNotificationMock)
+  }
+})
 jest.setMock('node-fetch', { default: fetch })
 process.env.CERT_PUBLIC_KEY_PATH = join(__dirname, './cert.key.pub')
 process.env.NODE_ENV = 'development'
